Moving external endian library to util/integer, moved int128/uint128 from util/numerics to util/integer as well.
