+++ hpx/plugins/parcelport/tcp/sender.hpp
@@ -26,7 +26,6 @@
 #include <boost/bind/protect.hpp>
 #include <boost/cstdint.hpp>
 #include <boost/enable_shared_from_this.hpp>
-#include <boost/integer/endian.hpp>
 #include <boost/noncopyable.hpp>
 #include <boost/shared_ptr.hpp>
 #include <boost/tuple/tuple.hpp>
