+++ hpx/runtime/parcelset/encode_parcels.hpp
@@ -13,9 +13,8 @@
 #include <hpx/runtime/parcelset/parcel_buffer.hpp>
 #include <hpx/runtime/serialization/serialize.hpp>
 #include <hpx/util/high_resolution_timer.hpp>
+#include <hpx/util/integer/endian.hpp>
 #include <hpx/traits/is_chunk_allocator.hpp>
-
-#include <boost/integer/endian.hpp>
 
 #include <memory>
