+++ hpx/runtime/parcelset/parcel_buffer.hpp
@@ -10,8 +10,8 @@
 #include <hpx/config.hpp>
 #include <hpx/runtime/serialization/serialization_chunk.hpp>
 #include <hpx/performance_counters/parcels/data_point.hpp>
+#include <hpx/util/integer/endian.hpp>
 
-#include <boost/integer/endian.hpp>
 #include <boost/atomic.hpp>
 
 #include <vector>
@@ -22,7 +22,7 @@ namespace hpx { namespace parcelset
     struct parcel_buffer
     {
         typedef std::pair<
-            boost::integer::ulittle32_t, boost::integer::ulittle32_t
+            util::integer::ulittle32_t, util::integer::ulittle32_t
         > count_chunks_type;
 
         typedef typename BufferType::allocator_type allocator_type;
@@ -84,15 +84,16 @@ namespace hpx { namespace parcelset
         BufferType data_;
         std::vector<ChunkType> chunks_;
 
-        typedef std::pair<boost::integer::ulittle64_t, boost::integer::ulittle64_t>
-            transmission_chunk_type;
+        typedef std::pair<
+            util::integer::ulittle64_t, util::integer::ulittle64_t
+        > transmission_chunk_type;
         std::vector<transmission_chunk_type> transmission_chunks_;
 
         // pair of (zero-copy, non-zero-copy) chunks
         count_chunks_type num_chunks_;
 
-        boost::integer::ulittle64_t size_;
-        boost::integer::ulittle64_t data_size_;
+        util::integer::ulittle64_t size_;
+        util::integer::ulittle64_t data_size_;
 
         /// Counters and their data containers.
         performance_counters::parcels::data_point data_point_;
