+++ src/util/integer/uint128.cpp
@@ -7,7 +7,7 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/hpx_fwd.hpp>
-#include <hpx/util/numerics/uint128.hpp>
+#include <hpx/util/integer/uint128.hpp>
 
 #include <memory>
 #include <cmath>
@@ -15,7 +15,7 @@
 
 // IMPLEMENTATION
 
-namespace hpx { namespace util { namespace numerics
+namespace hpx { namespace util { namespace integer
 {
     const char * uint128::toString (unsigned int radix) const throw () {
         if (!*this) return "0";
